feat:m:fixed and added basic authentication
