+++ apps/server/src/misc/crypto.js
@@ -1,0 +1,5 @@
+const crypto = require('crypto');
+
+const key = crypto.randomBytes(64).toString('hex');
+
+console.log(key);
