+++ apps/server/src/router/v1/auth.routes.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const mainController = require('../../controllers/main.controller');
+const mainController = require('../../controllers/auth.controller');
 
 router.post('/sign-in', mainController.signIn);
